+++ packages/cmscore/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @finsweet/attributes-cmscore
 
+## 1.7.1
+
+### Patch Changes
+
+- d8a4bf6f: Fixed invalid visibleCountFrom when there are no valid items
+
+## 1.7.0
+
+### Minor Changes
+
+- ef56d454: Added support for new `CMSList.visibleCountFrom` and `CMSList.visibleCountTo`.
+
 ## 1.6.10
 
 ### Patch Changes